Fix searching for posts made this month
Searching for ([style="white-space: nowrap; text-align: center;"]) returns other elements that do not contain what we intend to find, which is the current date. In the spirit of asking for forgiveness I have checked to see if the current element in the for loop does contain a valid string; if it doesn't, we move on to the next one.